@@ -47,7 +47,7 @@
     "p-series": "2.1.0",
     "plop": "^3.0.5",
     "pretty-ms": "7.0.1",
-    "react": "17.0.2",
+    "react": "18.2.0",
     "tailchat-server-sdk": "^0.0.12",
     "tailchat-shared": "workspace:*",
     "yargs": "^17.4.0"
@@ -57,7 +57,7 @@
     "@types/inquirer": "^8.2.1",
     "@types/lodash": "^4.14.170",
     "@types/node": "16.11.7",
-    "@types/react": "^17.0.39",
+    "@types/react": "18.0.20",
     "@types/yargs": "^17.0.10",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.3"
