@@ -39,16 +39,16 @@
     "@storybook/react": "^6.4.22",
     "@storybook/testing-library": "^0.0.11",
     "@types/lodash": "^4.14.170",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.8",
+    "@types/react": "18.0.20",
+    "@types/react-dom": "18.0.6",
     "babel-loader": "^8.2.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "typescript": "^4.5.2",
     "webpack": "^5.72.0"
   },
   "peerDependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   }
 }
