@@ -17,12 +17,12 @@
     "filesize": "^8.0.7",
     "flatted": "^3.2.4",
     "formik": "^2.2.9",
-    "i18next": "^20.3.2",
-    "i18next-http-backend": "^1.2.6",
+    "i18next": "^21.9.2",
+    "i18next-http-backend": "^1.4.1",
     "lodash": "^4.17.21",
-    "react-i18next": "^11.15.1",
+    "react-i18next": "^11.18.6",
     "react-native-storage": "npm:@trpgengine/react-native-storage@^1.0.1",
-    "react-redux": "^7.2.6",
+    "react-redux": "^8.0.2",
     "regenerator-runtime": "^0.13.9",
     "socket.io-client": "^4.1.2",
     "url-regex": "^5.0.0",
@@ -31,12 +31,11 @@
   "devDependencies": {
     "@types/crc": "^3.4.0",
     "@types/lodash": "^4.14.170",
-    "@types/react": "^17.0.39",
-    "@types/react-redux": "^7.1.24",
-    "react": "17.0.2"
+    "@types/react": "18.0.20",
+    "react": "18.2.0"
   },
   "peerDependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   }
 }
