@@ -42,14 +42,14 @@
     "p-min-delay": "^4.0.0",
     "qs": "^6.10.3",
     "rc-tree": "^5.3.6",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-easy-crop": "^3.5.2",
     "react-helmet": "^6.1.0",
     "react-markdown": "6",
     "react-mentions": "^4.3.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router": "^6.4.0",
+    "react-router-dom": "^6.4.0",
     "react-split": "^2.0.14",
     "react-transition-group": "^4.4.2",
     "react-use-gesture": "^9.1.3",
@@ -72,19 +72,19 @@
     "@types/emoji-mart": "^3.0.8",
     "@types/fs-extra": "^9.0.13",
     "@types/is-hotkey": "^0.1.5",
-    "@types/jest": "^27.0.3",
+    "@types/jest": "^29.0.3",
     "@types/loadable__component": "^5.13.4",
     "@types/lodash": "^4.14.170",
     "@types/mini-css-extract-plugin": "^1.4.3",
     "@types/minimatch": "^3.0.5",
     "@types/node": "^15.12.5",
     "@types/qs": "^6.9.7",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.8",
+    "@types/react": "18.0.20",
+    "@types/react-dom": "18.0.6",
     "@types/react-helmet": "^6.1.2",
     "@types/react-mentions": "^4.1.5",
-    "@types/react-router": "^5.1.15",
-    "@types/react-router-dom": "^5.1.7",
+    "@types/react-router": "^5.1.19",
+    "@types/react-router-dom": "^5.3.3",
     "@types/react-transition-group": "^4.4.2",
     "@types/react-virtualized": "^9.21.14",
     "@types/react-virtualized-auto-sizer": "^1.0.1",
