@@ -11,7 +11,7 @@
   },
   "devDependencies": {
     "@types/react-highlight": "^0.12.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   }
 }
