@@ -8,6 +8,6 @@
   },
   "devDependencies": {
     "@types/react-canvas-draw": "^1.1.1",
-    "react": "17.0.2"
+    "react": "18.2.0"
   }
 }
