@@ -8,6 +8,6 @@
     "html-react-parser": "^1.4.5"
   },
   "devDependencies": {
-    "react": "17.0.2"
+    "react": "18.2.0"
   }
 }
