@@ -8,8 +8,8 @@
     "tailchat-meeting-sdk": "1.2.4"
   },
   "devDependencies": {
-    "@types/react": "^17.0.38",
+    "@types/react": "18.0.20",
     "@types/react-router": "^5.1.18",
-    "react": "17.0.2"
+    "react": "18.2.0"
   }
 }
