@@ -4,7 +4,7 @@
   "version": "0.0.0",
   "private": true,
   "devDependencies": {
-    "@types/react": "^17.0.38",
-    "react": "17.0.2"
+    "@types/react": "18.0.20",
+    "react": "18.2.0"
   }
 }
