@@ -1,17 +1,16 @@
-import { useHistory } from 'react-router';
+import { NavigateOptions, To, useNavigate } from 'react-router';
 
 export interface QuickActionContext {
-  navigate: (url: string) => void;
+  navigate: (to: To, options?: NavigateOptions) => void;
 }
 
 /**
  * 快速切换操作上下文信息
  */
 export function useQuickSwitcherActionContext(): QuickActionContext {
-  const history = useHistory();
+  const navigate = useNavigate();
+
   return {
-    navigate: (url) => {
-      history.push(url);
-    },
+    navigate,
   };
 }
