@@ -3,7 +3,7 @@ import { GroupDetail } from '@/components/modals/GroupDetail';
 import { CreateGroupInvite } from '@/components/modals/CreateGroupInvite';
 import React from 'react';
 import { useCallback } from 'react';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router';
 import { quitGroup, showAlert, t, useIsGroupOwner } from 'tailchat-shared';
 import { useLocationNav } from '@/hooks/useHistoryNav';
 
@@ -12,7 +12,7 @@ import { useLocationNav } from '@/hooks/useHistoryNav';
  */
 export function useGroupHeaderAction(groupId: string) {
   const isOwner = useIsGroupOwner(groupId);
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const handleShowGroupDetail = useCallback(() => {
     const key = openModal(
@@ -36,7 +36,9 @@ export function useGroupHeaderAction(groupId: string) {
         : t('确定要退出群组么?'),
       async onConfirm() {
         await quitGroup(groupId);
-        history.replace('/main'); // 返回到主页
+        navigate('/main', {
+          replace: true,
+        }); // 返回到主页
       },
     });
   }, [groupId, isOwner]);
