@@ -1,11 +1,24 @@
 import { useRecordMeasure } from '@/utils/measure-helper';
 import React from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 import { MainProvider } from '../Main/Provider';
 import { t } from 'tailchat-shared';
 import { PersonalConverse } from '../Main/Content/Personal/Converse';
 import { GroupPanelRoute } from '../Main/Content/Group/Panel';
 import { GroupDetail } from '@/components/modals/GroupDetail';
+import { useParams } from 'react-router';
+import { NotFound } from '@/components/NotFound';
+
+const GroupDetailRoute = React.memo(() => {
+  const { groupId } = useParams<{ groupId: string }>();
+
+  if (!groupId) {
+    return <NotFound />;
+  }
+
+  return <GroupDetail groupId={groupId} onClose={() => {}} />;
+});
+GroupDetailRoute.displayName = 'GroupDetailRoute';
 
 const PanelRoute: React.FC = React.memo(() => {
   useRecordMeasure('AppRouteRenderStart');
@@ -13,30 +26,19 @@ const PanelRoute: React.FC = React.memo(() => {
   return (
     <div className="flex h-full bg-content-light dark:bg-content-dark">
       <MainProvider>
-        <Switch>
+        <Routes>
           <Route
-            exact={true}
-            path="/panel/personal/converse/:converseId"
-            component={PersonalConverse}
+            path="/personal/converse/:converseId"
+            element={<PersonalConverse />}
           />
+          <Route path="/group/:groupId/detail" element={<GroupDetailRoute />} />
           <Route
-            exact={true}
-            path="/panel/group/:groupId/detail"
-            render={(props) => (
-              <GroupDetail
-                groupId={props.match.params.groupId}
-                onClose={() => {}}
-              />
-            )}
-          />
-          <Route
-            exact={true}
-            path="/panel/group/:groupId/:panelId"
-            component={GroupPanelRoute}
+            path="/group/:groupId/:panelId"
+            element={<GroupPanelRoute />}
           />
 
-          <Route>{t('未知的面板')}</Route>
-        </Switch>
+          <Route path="/*" element={t('未知的面板')} />
+        </Routes>
       </MainProvider>
     </div>
   );
