@@ -31,12 +31,13 @@ export const App: React.FC = React.memo(() => {
 
       <Box>
         <Tabs flexDirection="column" onChange={() => {}}>
-          <Tab name="tab1">
+          {/* 因为react版本问题暂时注释 */}
+          {/* <Tab name="tab1">
             <Text>Foo</Text>
           </Tab>
           <Tab name="tab2">
             <Text>Bar</Text>
-          </Tab>
+          </Tab> */}
         </Tabs>
       </Box>
     </Box>
