@@ -1,7 +1,7 @@
-import React from 'react';
+import React, { PropsWithChildren } from 'react';
 import styles from './index.module.less';
 
-export const Highlight: React.FC = React.memo((props) => {
+export const Highlight: React.FC<PropsWithChildren> = React.memo((props) => {
   return <span className={styles.highLight}>{props.children}</span>;
 });
 Highlight.displayName = 'Highlight';
