@@ -1,5 +1,4 @@
 import { useParams } from 'react-router';
-import { GroupPanel, useGroupPanelInfo } from 'tailchat-shared';
 
 /**
  * 获取群组面板的参数
@@ -8,7 +7,7 @@ export function useGroupPanelParams(): {
   groupId: string;
   panelId: string;
 } {
-  const { groupId, panelId } = useParams<{
+  const { groupId = '', panelId = '' } = useParams<{
     groupId: string;
     panelId: string;
   }>();
