@@ -1,13 +1,14 @@
+import { NotFound } from '@/components/NotFound';
 import { ConversePanel } from '@/components/Panel/personal/ConversePanel';
 import React from 'react';
 import { useParams } from 'react-router';
 
-interface UserConversePanelParams {
-  converseId: string;
-}
-
 export const PersonalConverse: React.FC = React.memo(() => {
-  const params = useParams<UserConversePanelParams>();
+  const params = useParams<{ converseId: string }>();
+
+  if (!params.converseId) {
+    return <NotFound />;
+  }
 
   return <ConversePanel converseId={params.converseId} />;
 });
